devices/containers/demo-temp-sensor/requirements.txt
azure-iot-device
numpy